Делает загрузку комментариев часть 2 (module8-task2)
